// examples/chatbot/src/stateMachines/machines/query/constants/StateName.enum.ts
export enum StateName {
  IdentifyQueryCategoryOpenAiPrepare = 'IdentifyQueryCategoryOpenAiPrepare',
  IdentifyQueryCategoryOpenAi = 'IdentifyQueryCategoryOpenAi',
  QueryVectorCreatePinecone = 'QueryVectorCreatePinecone',
  QueryVectorSearchPinecone = 'QueryVectorSearchPinecone',
  QueryVectorSearchPineconeActions = 'QueryVectorSearchPineconeActions',
  QueryVectorSearchPineconeMemories = 'QueryVectorSearchPineconeMemories',
  ActionOrMemoryBranchingFail = 'ActionOrMemoryBranchingFail',
  VectorSearchResultsProcess = 'VectorSearchResultsProcess',
  ContextCollectRouter = 'ContextCollectRouter',
  ContextCollectResources = 'ContextCollectResources',
  ContextCollectAction = 'ContextCollectAction',
  ContextResourcesToContextString = 'ContextResourcesToContextString',
  SystemMessageOpenAiPrepareRouter = 'SystemMessageOpenAiPrepareRouter',
  SystemMessageMemoriesOpenAiPrepare = 'SystemMessageMemoriesOpenAiPrepare',
  SystemMessageActionsOpenAiPrepare = 'SystemMessageActionsOpenAiPrepare',
  QuestionAdd = 'QuestionAdd',
  ActionOrMemoryBranching = 'ActionOrMemoryBranching',
  ActionOpenAiPrepare = 'ActionOpenAiPrepare',
  GetConversation = 'GetConversation',
  ActionOpenAi = 'ActionOpenAi',
  ExecuteActionStateMachine = 'ExecuteActionStateMachine',
  ChildStateMachineSucceeded = 'ChildStateMachineSucceeded',
  ConversationOpenAiPrepare = 'ConversationOpenAiPrepare',
  ConversationOpenAi = 'ConversationOpenAi',
  SaveQueryAnswer = 'SaveQueryAnswer',
  QueryFinished = 'QueryFinished',
  IsNewConversationRouter = 'IsNewConversationRouter',
  NewLatestConversationId = 'NewLatestConversationId',
  UpdateLatestConversationId = 'UpdateLatestConversationId',
  SelectConversationToContinue = 'SelectConversationToContinue',
  GetLatestConversationId = 'GetLatestConversationId',
  GetContextForConversationId = 'GetContextForConversationId',
  SaveContext = 'SaveContext',
}